Tricky vtable fold complete: 614 and 529, and isGuarding really returns u8
The last two private Tricky export-table spellings now reach the shared
TrickyCompanionInterface in main/dll/dll_00C4_tricky.h through TRICKY_INTERFACE,
finishing the 31-DLL fold.

614_KytesMum dropped KytesMumTrickyInterface (ten opaque pointers, the 0x28
sideCommandEnable and the 0x34 commandPlayBall), its two STATIC_ASSERTs and its
private macro. One of its two call sites had also open-coded the vtable cast
instead of using the macro; both now go through TRICKY_INTERFACE.

529 needed the shared declaration corrected first. It typed the 0x44 slot as
returning u8 where the interface said int, and that u8 is the truth: retail
wmwallcrawler_update follows the bctrl with clrlwi r0,r3,24 / cmplwi r0,0, which
is the u8 width -- folding onto the int spelling collapses that to a bare
cmpwi r3,0 and loses an instruction. The two TrickyGuard DLLs were already
compensating for the wrong header with an explicit (u8) cast at all three of
their call sites, which is the same evidence from the other direction. So the
field is now u8, the three compensating casts are gone, and 529 folds byte-clean.

Byte-verified: all 991 source objects are identical before and after, so the
narrow and the cast removals are byte-neutral everywhere. ninja all_source and
ninja EXIT=0, dtk shasum main.dol OK, report.json fuzzy 99.77637 /
complete_units 881, both unchanged.

+++ include/main/dll/dll_00C4_tricky.h
@@ -35,7 +35,7 @@ typedef struct TrickyCompanionInterface
     int (*requestMoveToObject)(GameObject* tricky, GameObject* target);
     void (*requestRecall)(GameObject* tricky);
     u8 (*isPlayingBall)(GameObject* tricky);
-    int (*isGuarding)(GameObject* tricky);
+    u8 (*isGuarding)(GameObject* tricky);
     int (*getCurrentCommandType)(GameObject* tricky, int* commandType);
 } TrickyCompanionInterface;
